Refactor to minimise merge issues
