+++ app/services/return-logs/setup/initiate-session.service.js
@@ -7,8 +7,8 @@
 
 const { ref } = require('objection')
 
-const ReturnLogModel = require('../../models/return-log.model.js')
-const SessionModel = require('../../models/session.model.js')
+const ReturnLogModel = require('../../../models/return-log.model.js')
+const SessionModel = require('../../../models/session.model.js')
 
 /**
  * Initiates the session record used for setting up a new return log edit journey
